Don't use screen to run cAdvisor in the background.
Some distros don't have screen. Here we wait for cAdvisor to come up
before continuing.
